+++ test/test_tenancy_scopes.py
@@ -28,9 +28,7 @@ STUB_LLM = cast(LLMTool, object())
 
 
 def _config(tmp_path) -> Config:
-    config = Config.in_memory(
-        tool_config=ToolConfig(path_config=PathConfig(working_directory=tmp_path))
-    )
+    config = Config.in_memory(tool_config=ToolConfig(path_config=PathConfig()))
     config.tool_config.vector_store.backend = VectorStoreBackend.NONE
     return config